pyproject.toml
[tool.poetry]
name = "ai-resume-screener"
version = "0.1.0"
description = ""
authors = ["Chris Ragobeer <chris.ragobeer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
langchain = "^0.2.16"
llama-index = "^0.11.7"
llama-index-embeddings-ollama = "^0.3.0"
llama-index-llms-ollama = "^0.3.1"
qdrant-client = "^1.11.1"
llama-index-vector-stores-qdrant = "^0.3.0"
unstructured = "^0.15.10"
lxml = "^5.3.0"
nltk = "^3.9.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
